requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.8.2
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
httpie==0.9.9
idna==2.6
markdown2==2.3.1
Pillow==5.0.0
Pygments==2.2.0
PyJWT==1.4.0
pytz==2018.3
requests==2.18.4
urllib3==1.22
